Fixed new loss, commented out arg parse section
I fixed the new_loss function, Jeff- I don't know remember exactly what your formulation for the actual loss is so you will have to put those in. however I made the norms.

I also commented out a the section of sams code that limits string parsing, as it was causeing problems with the model imput. With sam's code it passes a list not a string.